requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.4
pyarrow==14.0.2
pydantic==1.10.13
scikit-learn==1.3.2
lightgbm==4.1.0
optuna==3.5.0
mlflow==2.9.2
streamlit==1.30.0
ta==0.10.2
pytz==2023.3.post1
dateparser==1.2.0
joblib==1.3.2
rich==13.7.0
tqdm==4.66.2
matplotlib==3.8.2
